Changed "float" to "double" in UtenteService, UtenteServiceImpl and RivenditaServiceImpl

+++ src/main/java/it/unisa/c02/moneyart/gestione/utente/service/UtenteService.java
@@ -31,7 +31,7 @@ public interface UtenteService {
 
   boolean withdraw(Utente utente, float amount);
 
-  float getBalance(Utente utente);
+  double getBalance(Utente utente);
 
   boolean transfer(Utente sender, Utente receiver, float amount);
